requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
django-cors-headers==4.9.0
Pillow==11.3.0
mysqlclient==2.2.7
django-allauth==65.12.0
requests==2.32.5
PyJWT==2.10.1
python-dotenv==1.0.0

# Платежные системы
yookassa==3.4.1

# Службы доставки
# russianpost - используем собственную реализацию через requests
whitenoise==6.7.0
